# openapi.yaml
openapi: 3.0.3
info:
  title: Hierarchical Agent System API
  description: |
    REST API for managing hierarchical agent teams, configurations, and task executions.
    
    This API provides comprehensive functionality for:
    - Creating and managing hierarchical agent teams
    - Discovering and searching agent libraries
    - Validating and optimizing team configurations
    - Executing tasks with real-time monitoring
    - Performance analytics and team evaluation
    
    ## Authentication
    Currently operates without authentication for development. Production deployments support:
    - API Key authentication via `X-API-Key` header
    - JWT Bearer tokens via `Authorization` header
    
    ## Rate Limiting
    - Default: 100 requests/minute per IP
    - Authenticated: 1000 requests/minute per API key
    - Execution endpoints: 10 concurrent executions per team
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format with structured error responses.
    
  version: 1.0.0
  contact:
    name: Hierarchical Agents Support
    url: https://github.com/your-org/hierarchical-agents
    email: support@hierarchical-agents.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.hierarchical-agents.com
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {}
...
paths:
...
  /api/v1/teams/:
    post:
      tags: [teams]
      summary: Create Team
      description: Create a new hierarchical agent team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreateRequest'
            examples:
              research_team:
                summary: Research Team Example
                value:
                  name: "Advanced Research Team"
                  description: "Multi-disciplinary research and analysis team"
                  config:
                    coordinator:
                      routing_strategy: "capability_based"
                      prompts:
                        system_prompt:
                          template: "You are a coordinator managing specialized research teams..."
                    teams:
                      - name: "Research Team"
                        supervisor:
                          name: "Research Supervisor"
                          prompts:
                            system_prompt:
                              template: "You supervise research tasks and coordinate with multiple researchers..."
                        workers:
                          - name: "Senior Researcher"
                            capabilities: ["research", "analysis", "data_processing"]
                            tools: ["web_search", "calculator", "file_reader"]
      responses:
        '201':
          description: Team created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [teams]
      summary: List Teams
      description: List all teams with pagination and filtering
      parameters:
        - name: limit
          in: query
          description: Maximum number of teams to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of teams to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: status
          in: query
          description: Filter by team status
          schema:
            type: string
            enum: [active, inactive, error, maintenance]
        - name: search
          in: query
          description: Search in team names and descriptions
          schema:
            type: string
            maxLength: 200
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
  /api/v1/teams/{team_id}:
    parameters:
      - $ref: '#/components/parameters/TeamId'
    
    get:
      tags: [teams]
      summary: Get Team
      description: Get detailed information about a specific team
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [teams]
      summary: Update Team
      description: Update an existing team configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamUpdateRequest'
      responses:
        '200':
          description: Team updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags: [teams]
      summary: Delete Team
      description: Delete a team and stop all active executions
      parameters:
        - name: force
          in: query
          description: Force delete even with active executions
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Team deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDeleteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/teams/{team_id}/status:
    parameters:
      - $ref: '#/components/parameters/TeamId'
    
    get:
      tags: [teams]
      summary: Get Team Status
      description: Get real-time team status and performance metrics
      responses:
        '200':
          description: Team status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # Agent Library
...
  /api/v1/executions/{team_id}/execute:
    parameters:
      - $ref: '#/components/parameters/TeamId'
    
    post:
      tags: [executions]
      summary: Execute Task
      description: Execute a task using a hierarchical team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequest'
            examples:
              research_task:
                summary: Research Task
                value:
                  input_text: "Research the latest developments in quantum computing and provide a comprehensive report"
                  parameters:
                    priority: "high"
                    format: "detailed_report"
                    target_audience: "technical executives"
                  timeout_seconds: 300
              content_creation:
                summary: Content Creation Task
                value:
                  input_text: "Write a blog post about sustainable energy trends"
                  parameters:
                    word_count: 1500
                    tone: "informative yet engaging"
                    target_audience: "general public"
                  timeout_seconds: 600
      responses:
        '202':
          description: Task execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Team busy with maximum concurrent executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/executions/{execution_id}:
    parameters:
      - $ref: '#/components/parameters/ExecutionId'
    
    get:
      tags: [executions]
      summary: Get Execution Status
      description: Get execution status and results
      responses:
        '200':
          description: Execution details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/executions/:
    get:
      tags: [executions]
      summary: List Executions
      description: List executions with filtering and pagination
      parameters:
        - name: team_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort
          in: query
          schema:
            type: string
            enum: [started_at, -started_at, duration, -duration, cost, -cost]
            default: "-started_at"
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionListResponse'
  /api/v1/executions/{execution_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/ExecutionId'
    
    post:
      tags: [executions]
      summary: Cancel Execution
      description: Cancel a running execution
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionCancelRequest'
      responses:
        '200':
          description: Execution cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionCancelResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TeamId:
      name: team_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]+$'
      description: Unique team identifier
      example: team_abc123
...
    ExecutionId:
      name: execution_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]+$'
      description: Unique execution identifier
      example: exec_xyz789
...
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid request data"
              details:
                field: "config.teams"
                issue: "At least one team is required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Team not found"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "CONFLICT"
              message: "Team with this name already exists"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Configuration validation failed"
              details:
                issues:
                  - severity: "error"
                    path: "teams[0].workers"
                    message: "At least one worker is required"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Request validation failed"
            details:
              type: object
              description: Additional error context
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
              example: "req_abc123"

    # Team Models
    TeamCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9_\s-]+$'
          example: "Advanced Research Team"
        description:
          type: string
          maxLength: 500
          example: "Multi-disciplinary research and analysis team"
        config_data:
          $ref: '#/components/schemas/HierarchicalAgentConfiguration'
          description: "Complete hierarchical team configuration"
        config_file_path:
          type: string
          description: "Path to YAML configuration file"
          example: "configs/examples/hierarchical/research_team.yml"
        template_id:
          type: string
          description: "ID of template to use for team creation"
          example: "research_pipeline"
    TeamUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        config:
          $ref: '#/components/schemas/TeamConfiguration'
    TeamResponse:
      type: object
      properties:
        team_id:
          type: string
          example: "team_abc123"
        name:
          type: string
          example: "Advanced Research Team"
        description:
          type: string
          example: "Multi-disciplinary research and analysis team"
        status:
          type: string
          enum: [active, inactive, error, maintenance]
          example: active
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        config:
          $ref: '#/components/schemas/TeamConfiguration'
        performance_metrics:
          $ref: '#/components/schemas/PerformanceMetrics'
    TeamListResponse:
      type: object
      properties:
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamSummary'
        total:
          type: integer
          example: 42
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
    TeamSummary:
      type: object
      properties:
        team_id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive, error, maintenance]
        created_at:
          type: string
          format: date-time
        performance_metrics:
          $ref: '#/components/schemas/PerformanceMetrics'
    TeamDeleteResponse:
      type: object
      properties:
        team_id:
          type: string
        message:
          type: string
          example: "Team deleted successfully"
        cancelled_executions:
          type: integer
          example: 2
    TeamStatusResponse:
      type: object
      properties:
        team_id:
          type: string
        status:
          type: string
          enum: [active, inactive, error, maintenance]
        health_score:
          type: number
          minimum: 0
          maximum: 1
          example: 0.87
        active_executions:
          type: integer
          example: 3
        last_execution:
          type: string
          format: date-time
        performance_metrics:
          $ref: '#/components/schemas/PerformanceMetrics'
        resource_usage:
          type: object
          properties:
            cpu_usage:
              type: number
              example: 0.45
            memory_usage:
              type: number
              example: 0.67
            token_usage_24h:
              type: integer
              example: 15420
    PerformanceMetrics:
      type: object
      properties:
        total_executions:
          type: integer
          example: 156
        success_rate:
          type: number
          minimum: 0
          maximum: 1
          example: 0.94
        avg_response_time:
          type: number
          example: 12.3
        avg_cost_per_execution:
          type: number
          example: 0.05

    # Hierarchical Agent Configuration
    HierarchicalAgentConfiguration:
      type: object
      required: [team, coordinator, teams]
      properties:
        team:
          $ref: '#/components/schemas/TeamInfo'
        coordinator:
          $ref: '#/components/schemas/CoordinatorConfig'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamConfig'
          minItems: 1
        memory:
          $ref: '#/components/schemas/HierarchicalMemoryConfig'
        coordination:
          $ref: '#/components/schemas/CoordinationConfig'
        performance:
          $ref: '#/components/schemas/PerformanceConfig'
        deployment:
          $ref: '#/components/schemas/DeploymentConfig'
        runtime:
          $ref: '#/components/schemas/HierarchicalRuntimeConfig'
    TeamInfo:
      type: object
      required: [name, description, type]
      properties:
        name:
          type: string
          example: "Research Pipeline Team"
        description:
          type: string
          example: "Hierarchical team for comprehensive research"
        version:
          type: string
          example: "1.0.0"
        type:
          type: string
          enum: [hierarchical]
          example: "hierarchical"

    # Legacy Team Configuration (kept for compatibility)
    TeamConfiguration:
      type: object
      required: [coordinator, teams]
      properties:
        coordinator:
          $ref: '#/components/schemas/CoordinatorConfig'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamConfig'
          minItems: 1
        performance:
          $ref: '#/components/schemas/PerformanceConfig'
    CoordinatorConfig:
      type: object
      required: [name, description, llm, prompts]
      properties:
        name:
          type: string
          example: "Research Coordinator"
        description:
          type: string
          example: "Coordinates research tasks across specialized teams"
        llm:
          $ref: '#/components/schemas/LLMConfig'
        prompts:
          $ref: '#/components/schemas/CoordinatorPromptsConfig'
        routing:
          $ref: '#/components/schemas/RoutingConfig'
        memory:
          $ref: '#/components/schemas/MemoryConfig'
    CoordinatorPromptsConfig:
      type: object
      required: [system_prompt, decision_prompt]
      properties:
        system_prompt:
          $ref: '#/components/schemas/PromptTemplate'
        decision_prompt:
          $ref: '#/components/schemas/PromptTemplate'
        coordination_prompt:
          $ref: '#/components/schemas/PromptTemplate'
    RoutingConfig:
      type: object
      properties:
        strategy:
          type: string
          enum: [keyword_based, llm_based, rule_based, hybrid]
          default: hybrid
        fallback_team:
          type: string
        fallback_worker:
          type: string
        confidence_threshold:
          type: number
          default: 0.7
        max_decision_time:
          type: integer
          default: 30
    TeamConfig:
      type: object
      required: [name, supervisor, workers]
      properties:
        name:
          type: string
          example: "Research Team"
        supervisor:
          $ref: '#/components/schemas/SupervisorConfig'
        workers:
          type: array
          items:
            $ref: '#/components/schemas/WorkerConfig'
          minItems: 1
...
    LLMConfig:
      type: object
      required: [provider, model]
      properties:
        provider:
          type: string
          enum: [openai, anthropic, google, groq]
        model:
          type: string
          example: "gpt-4o"
        temperature:
          type: number
          minimum: 0
          maximum: 2
          example: 0.7
        max_tokens:
          type: integer
          minimum: 1
          example: 2048
        timeout:
          type: integer
          minimum: 1
          example: 30
        api_key_env:
          type: string
          example: "OPENAI_API_KEY"
          description: "Environment variable name containing the API key"
...
    MemoryConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        type:
          type: string
          enum: [semantic, episodic, procedural]
        retention_policy:
          type: string
          enum: [session, persistent, limited]
    PerformanceConfig:
      type: object
      properties:
        monitoring_enabled:
          type: boolean
          default: true
        alerts:
          type: array
          items:
            type: object
            properties:
              metric:
                type: string
              threshold:
                type: number
              action:
                type: string

    # Agent Models
...
    ExecutionRequest:
      type: object
      required: [input_text]
      properties:
        input_text:
          type: string
          minLength: 1
          maxLength: 10000
          example: "Research the latest developments in quantum computing"
        team_id:
          type: string
          description: "Specific team to execute with (optional)"
        worker_id:
          type: string
          description: "Specific worker to execute with (optional)"
        timeout_seconds:
          type: integer
          minimum: 10
          maximum: 3600
          default: 120
          description: "Execution timeout in seconds"
        parameters:
          type: object
          description: "Additional execution parameters"
          properties:
            priority:
              type: string
              enum: [low, medium, high, urgent]
            deadline:
              type: string
              format: date-time
            format:
              type: string
              example: "detailed_report"
            audience:
              type: string
              example: "technical executives"
            constraints:
              type: array
              items:
                type: string
          additionalProperties: true
    ExecutionResponse:
      type: object
      properties:
        execution_id:
          type: string
        team_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        started_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
          format: date-time
        progress:
          $ref: '#/components/schemas/ExecutionProgress'
    ExecutionProgress:
      type: object
      properties:
        percentage:
          type: number
          minimum: 0
          maximum: 100
        current_step:
          type: string
        steps_completed:
          type: integer
        total_steps:
          type: integer
        eta_seconds:
          type: number
    ExecutionStatusResponse:
      type: object
      properties:
        execution_id:
          type: string
        team_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_seconds:
          type: number
        progress:
          $ref: '#/components/schemas/ExecutionProgress'
        result:
          $ref: '#/components/schemas/ExecutionResult'
        error:
          $ref: '#/components/schemas/ExecutionError'
        performance_metrics:
          type: object
          properties:
            response_time:
              type: number
            accuracy_score:
              type: number
            team_coordination_score:
              type: number
            resource_efficiency:
              type: number
    ExecutionResult:
      type: object
      properties:
        success:
          type: boolean
        output:
          type: string
        intermediate_outputs:
          type: array
          items:
            type: object
        metadata:
          type: object
          properties:
            tokens_used:
              type: integer
            agents_involved:
              type: array
              items:
                type: string
            cost_estimate:
              type: string
            execution_path:
              type: array
              items:
                type: object
    ExecutionError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        recovery_suggestions:
          type: array
          items:
            type: string
    ExecutionListResponse:
      type: object
      properties:
        executions:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionSummary'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    ExecutionSummary:
      type: object
      properties:
        execution_id:
          type: string
        team_id:
          type: string
        status:
          type: string
        started_at:
          type: string
          format: date-time
        duration_seconds:
          type: number
        success:
          type: boolean
    ExecutionCancelRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 200
        force:
          type: boolean
          default: false
    ExecutionCancelResponse:
      type: object
      properties:
        execution_id:
          type: string
        status:
          type: string
        cancelled_at:
          type: string
          format: date-time
        message:
          type: string
...